Added some more documentation to improve workflow
