+++ package.json
@@ -1,7 +1,6 @@
 {
 	"name": "nerdee",
 	"version": "0.1.0",
-
 	"private": true,
 	"dependencies": {
 		"@testing-library/jest-dom": "^5.11.4",
@@ -11,9 +10,11 @@
 		"@types/node": "^12.0.0",
 		"@types/react": "^17.0.0",
 		"@types/react-dom": "^17.0.0",
+		"@types/react-router-dom": "^5.1.6",
 		"node-sass": "^5.0.0",
 		"react": "^17.0.2",
 		"react-dom": "^17.0.2",
+		"react-router-dom": "^5.2.0",
 		"react-scripts": "4.0.3",
 		"typescript": "^4.1.2",
 		"web-vitals": "^1.0.1"
