--- src/components/index.ts
@@ -1,2 +1,0 @@
-export * from './items/Navbar';
-export * from './items/Preloader';
