--- src/components/items/Preloader.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-interface Props {}
-
-export const Preloader: React.FC<Props> = () => {
-	return <div /*id='preloader'*/ />;
-};
